Triage_System/Live_hosting_Project/requirements.txt
flask
python-dotenv
psycopg2-binary
apscheduler
gunicorn
